% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageDepth.R
\name{coverageDepth}
\alias{coverageDepth}
\title{extract coverage depth for gene level or transcript level}
\usage{
coverageDepth(RPFs, RNAs, gtf, level = c("tx", "gene"), bestpsite = 13,
  readsLen = c(28, 29), region = "cds", ...)
}
\arguments{
\item{RPFs}{Bam file names of RPFs.}

\item{RNAs}{Bam file names of RNAseq.}

\item{gtf}{GTF file name for annotation.}

\item{level}{transcript or gene level.}

\item{bestpsite}{P site postion.}

\item{readsLen}{reads length to keep.}

\item{region}{annotation region. It could be "cds", "utr5", "utr3" or "exon".}

\item{...}{parameters pass to \link[Rsubread:featureCounts]{featureCounts}}
}
\value{
a list with reads counts.
}
\description{
Calculate the coverage depth for gene level or transcript level.
Coverage for RPFs will be the best P site coverage.
Coverage for RNAs will be the coverage for 5'end of reads.
}
\examples{
\dontrun{
path <- system.file("extdata", package="ribosomeProfilingQC")
RPFs <- dir(path, "RPF.*?\\\\.[12].bam$", full.names=TRUE)
RNAs <- dir(path, "mRNA.*?\\\\.[12].bam$", full.names=TRUE)
gtf <- file.path(path, "Danio_rerio.GRCz10.91.chr1.gtf.gz")
cvgs <- coverageDepth(RPFs, RNAs, gtf, level="tx")
}
}
