% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvgd.R
\docType{class}
\name{cvgd-class}
\alias{cvgd-class}
\alias{cvgd}
\alias{$,cvgd-method}
\alias{$<-,cvgd-method}
\alias{[[,cvgd,ANY,ANY-method}
\alias{[[<-,cvgd,ANY,ANY,ANY-method}
\alias{show,cvgd-method}
\title{Class \code{"cvgd"}}
\usage{
cvgd(...)

\S4method{$}{cvgd}(x, name)

\S4method{$}{cvgd}(x, name) <- value

\S4method{[[}{cvgd,ANY,ANY}(x, i, j, ..., exact = TRUE)

\S4method{[[}{cvgd,ANY,ANY,ANY}(x, i, j, ...) <- value

\S4method{show}{cvgd}(object)
}
\arguments{
\item{\dots}{Each argument in \dots becomes an slot in the new \code{"cvgd"}-class.}

\item{x}{cvgd object.}

\item{name}{A literal character string or a name (possibly backtick quoted).}

\item{value}{value to replace.}

\item{i, j}{indices specifying elements to extract or replace.}

\item{exact}{see \link[base]{Extract}}

\item{object}{cvgd object.}
}
\description{
An object of class \code{"cvgd"}
represents output of \code{coverageDepth}.
}
\section{Slots}{

\describe{
\item{\code{coverage}}{\code{"list"}, list of \link[IRanges:AtomicList]{CompressedRleList},
specify the coverage of features of each sample.}

\item{\code{granges}}{\link[GenomicRanges:GRangesList-class]{CompressedGRangesList},
specify the features.}
}}

\examples{
cvgd()
}
