% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBar.R
\name{ggBar}
\alias{ggBar}
\title{barplot by ggplot2}
\usage{
ggBar(height, fill = "gray80", draw = TRUE, xlab, ylab, postfix)
}
\arguments{
\item{height}{data for plot}

\item{fill, xlab, ylab}{parameters pass to ggplot.}

\item{draw}{plot or not}

\item{postfix}{Postfix of text labled in top of bar.}
}
\value{
ggplot object.
}
\description{
barplot with number in top.
}
\examples{
ribosomeProfilingQC:::ggBar(sample.int(100, 3))
}
