% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFrameDensity.R
\name{plotFrameDensity}
\alias{plotFrameDensity}
\title{Plot density for each reading frame}
\usage{
plotFrameDensity(
  reads,
  density = TRUE,
  col = c(Frame_0 = "#009E73", Frame_1 = "#D55E00", Frame_2 = "#0072B2")
)
}
\arguments{
\item{reads}{Output of \link{assignReadingFrame}}

\item{density}{Plot density or counts}

\item{col}{Colors for reading frames}
}
\value{
Reading frame density
}
\description{
Plot density for each reading frame.
}
\examples{
pcs <- readRDS(system.file("extdata", "samplePc.rds",
               package="ribosomeProfilingQC"))
plotFrameDensity(pcs)
}
