% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translationalEfficiency.R
\name{translationalEfficiency}
\alias{translationalEfficiency}
\title{Translational Efficiency}
\usage{
translationalEfficiency(
  x,
  window,
  RPFsampleOrder,
  mRNAsampleOrder,
  pseudocount = 1,
  log2 = FALSE,
  normByLibSize = FALSE
)
}
\arguments{
\item{x}{output of \link{getFPKM} or \link{normByRUVs}.
if window is set, it must be output of \link{coverageDepth}.}

\item{window}{numeric(1). window size for maximal counts.}

\item{RPFsampleOrder, mRNAsampleOrder}{sample order of RPFs and mRNAs. The parameters
are used to make sure that the order of RPFs and mRNAs in cvgs is corresponding samples.}

\item{pseudocount}{the number will be add to sum of reads count to avoid X/0.}

\item{log2}{Do log2 transform or not.}

\item{normByLibSize}{normlization by library size or not.
If window size is provied and normByLibSize is set to TRUE,
the coverage will be normalized by library size.}
}
\value{
a list with RPFs, mRNA levels and TE as a matrix with translational efficiency
}
\description{
calculate translational efficiency.
}
\examples{
\dontrun{
path <- system.file("extdata", package="ribosomeProfilingQC")
RPFs <- dir(path, "RPF.*?\\\\.[12].bam$", full.names=TRUE)
RNAs <- dir(path, "mRNA.*?\\\\.[12].bam$", full.names=TRUE)
gtf <- file.path(path, "Danio_rerio.GRCz10.91.chr1.gtf.gz")
cnts <- countReads(RPFs, RNAs, gtf, level="gene")
fpkm <- getFPKM(cnts)
te <- translationalEfficiency(fpkm)
}
}
