% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignReadingFrame.R
\name{assignReadingFrame}
\alias{assignReadingFrame}
\title{assign reading frame}
\usage{
assignReadingFrame(reads, CDS, txdb)
}
\arguments{
\item{reads}{output of \link{getPsiteCoordinates}}

\item{CDS}{output of \link{prepareCDS}}

\item{txdb}{an TxDb object. If it is set, assign reading frame for all reads.
Default missing, only assign rading frame for reads in CDS.}
}
\value{
reads with reading frame information
}
\description{
set reading frame for each reads in CDS region
}
\examples{
library(Rsamtools)
bamfilename <- system.file("extdata", "RPF.WT.1.bam",
                           package="ribosomeProfilingQC")
yieldSize <- 10000000
bamfile <- BamFile(bamfilename, yieldSize = yieldSize)
pc <- getPsiteCoordinates(bamfile, bestpsite=13)
pc.sub <- pc[pc$qwidth \%in\% c(29, 30)]
#library(GenomicFeatures)
library(BSgenome.Drerio.UCSC.danRer10)
#txdb <- makeTxDbFromGFF(system.file("extdata",
 #         "Danio_rerio.GRCz10.91.chr1.gtf.gz",
 #         package="ribosomeProfilingQC"),
 #         organism = "Danio rerio",
 #         chrominfo = seqinfo(Drerio)["chr1"],
 #         taxonomyId = 7955)
#CDS <- prepareCDS(txdb)
CDS <- readRDS(system.file("extdata", "CDS.rds",
                           package="ribosomeProfilingQC"))
pc.sub <- assignReadingFrame(pc.sub, CDS)
}
