% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePsite.R
\name{estimatePsite}
\alias{estimatePsite}
\alias{bestPsite}
\title{estimate P site position}
\usage{
estimatePsite(bamfile, CDS, genome)

bestPsite(psite)
}
\arguments{
\item{bamfile}{an BamFile object.}

\item{CDS}{output of \link{prepareCDS}}

\item{genome}{a BSgenome object.}

\item{psite}{output of estimatePsite function.}
}
\value{
A list with strart codon position, stop codon position from 5 end of the reads,
and summary of positions.

The Psite postion.
}
\description{
Estimate P site postion from a subset reads.
}
\examples{
library(Rsamtools)
bamfilename <- system.file("extdata", "RPF.chr1.bam",
                           package="ribosomeProfilingQC")
yieldSize <- 10000000
bamfile <- BamFile(bamfilename, yieldSize = yieldSize)
library(BSgenome.Drerio.UCSC.danRer10)
CDS <- readRDS(system.file("extdata", "sampleCDS.rds",
               package="ribosomeProfilingQC"))
psite <- estimatePsite(bamfile, CDS, Drerio)
bestPsite(psite)
}
\references{
1: Bazzini AA, Johnstone TG, Christiano R, Mackowiak SD, Obermayer B, Fleming ES,
Vejnar CE, Lee MT, Rajewsky N, Walther TC, Giraldez AJ. Identification of small
ORFs in vertebrates using ribosome footprinting and evolutionary conservation.
EMBO J. 2014 May 2;33(9):981-93. doi: 10.1002/embj.201488411. Epub 2014 Apr 4.
PubMed PMID: 24705786; PubMed Central PMCID: PMC4193932.
}
