% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCDS.R
\name{filterCDS}
\alias{filterCDS}
\title{filter CDS by size}
\usage{
filterCDS(CDS, sizeCutoff = 100L)
}
\arguments{
\item{CDS}{output of preparedCDS}

\item{sizeCutoff}{cutoff size for CDS.
If the size of CDS is less than the cutoff, it will be filtered out.}
}
\value{
an GRanges object with filtered CDS.
}
\description{
filter CDS by CDS size.
}
\examples{
cds <- readRDS(system.file("extdata", "sampleCDS.rds",
               package="ribosomeProfilingQC"))
filterCDS(cds)
}
