% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLOSS.R
\name{FLOSS}
\alias{FLOSS}
\title{Fragment Length Organization Similarity Score (FLOSS)}
\usage{
FLOSS(
  reads,
  ref,
  CDS,
  readLengths = c(26:34),
  level = c("tx", "gene"),
  draw = FALSE
)
}
\arguments{
\item{reads}{output of \link{getPsiteCoordinates}}

\item{ref}{refercence id list}

\item{CDS}{output of \link{prepareCDS}}

\item{readLengths}{read length used for calculation}

\item{level}{transcript or gene level}

\item{draw}{plot FLOSS vs total reads or not.}
}
\value{
a data frame with colnames as id, FLOSS, totalReads, wilcox.test.pval, cook's distance.
}
\description{
The FLOSS will be calculated from a histogram of read lengths for footprints on a transcript or reading frame.
}
\examples{
library(Rsamtools)
bamfilename <- system.file("extdata", "RPF.chr1.bam",
                           package="ribosomeProfilingQC")
yieldSize <- 10000000
bamfile <- BamFile(bamfilename, yieldSize = yieldSize)
pc <- getPsiteCoordinates(bamfile, bestpsite=13)
CDS <- readRDS(system.file("extdata", "sampleCDS.rds",
               package="ribosomeProfilingQC"))
set.seed(123)
ref <- sample(unique(CDS$gene_id), 100)
fl <- FLOSS(pc, ref, CDS, level="gene")
}
\references{
1: Ingolia NT, Brar GA, Stern-Ginossar N, Harris MS, Talhouarne GJ, Jackson SE,
Wills MR, Weissman JS. Ribosome profiling reveals pervasive translation outside
of annotated protein-coding genes. Cell Rep. 2014 Sep 11;8(5):1365-79. doi:
10.1016/j.celrep.2014.07.045. Epub 2014 Aug 21. PubMed PMID: 25159147; PubMed
Central PMCID: PMC4216110.
}
