% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPsiteCoordinates.R
\name{getPsiteCoordinates}
\alias{getPsiteCoordinates}
\title{get P site coordinates}
\usage{
getPsiteCoordinates(bamfile, bestpsite)
}
\arguments{
\item{bamfile}{an BamFile object.}

\item{bestpsite}{P site postion. See \link{estimatePsite}}
}
\value{
A GRanges object with qwidth metadata which indicates the width of reads.
}
\description{
read in all reads from a bamfile and extract P site coordinates.
}
\examples{
library(Rsamtools)
bamfilename <- system.file("extdata", "tophat2.danRer10.RPF.chr1.bam",
                           package="ribosomeProfilingQC")
yieldSize <- 10000000
bamfile <- BamFile(bamfilename, yieldSize = yieldSize)
pc <- getPsiteCoordinates(bamfile, bestpsite=13)
}
