% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAmotif.R
\name{PAmotif}
\alias{PAmotif}
\title{Metaplot of P site distribution}
\usage{
PAmotif(reads, genome, plot = TRUE)
}
\arguments{
\item{reads}{Output of \link{assignReadingFrame} or \link{shiftReadsByFrame}.}

\item{genome}{A BSgenome object.}

\item{plot}{Plot the motif or not.}
}
\value{
A \link[motifStack:pcm-class]{pcm} object
}
\description{
Metaplot of P site distribution in all the CDS aligned by
the start codon or stop codon.
}
\examples{
pcs <- readRDS(system.file("extdata", "samplePc.rds",
               package="ribosomeProfilingQC"))
library(BSgenome.Drerio.UCSC.danRer10)
#PAmotif(pcs, Drerio)

}
