% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTE.R
\name{plotTE}
\alias{plotTE}
\title{plot translational efficiency}
\usage{
plotTE(
  TE,
  sample,
  xaxis = c("mRNA", "RPFs"),
  removeZero = TRUE,
  log2 = TRUE,
  breaks.length = 50,
  ...
)
}
\arguments{
\item{TE}{output of \link{translationalEfficiency}}

\item{sample}{character(1). Sample name to plot.}

\item{xaxis}{what to plot for x-axis.}

\item{removeZero}{Remove the 0 values from plots.}

\item{log2}{do log2 transform or not.}

\item{breaks.length}{length of breaks for histogram.}

\item{...}{parameters pass to \link[graphics:plot]{plot}.}
}
\value{
invisible data.frame with x, y of points.
}
\description{
Scatterplot of RNA/RPFs level compared to the
translational efficiency.
}
\examples{
path <- system.file("extdata", package="ribosomeProfilingQC")
#RPFs <- dir(path, "RPF.*?\\.[12].bam$", full.names=TRUE)
#RNAs <- dir(path, "mRNA.*?\\.[12].bam$", full.names=TRUE)
#gtf <- file.path(path, "Danio_rerio.GRCz10.91.chr1.gtf.gz")
#cnts <- countReads(RPFs, RNAs, gtf, level="gene")
cnts <- readRDS(file.path(path, "cnts.rds"))
fpkm <- getFPKM(cnts)
te <- translationalEfficiency(fpkm)
plotTE(te, 1)
}
