% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTranscript.R
\name{plotTranscript}
\alias{plotTranscript}
\title{plot reads P site abundance for a specific transcript}
\usage{
plotTranscript(
  reads,
  tx_name,
  col = c(Frame_0 = "#009E73", Frame_1 = "#D55E00", Frame_2 = "#0072B2")
)
}
\arguments{
\item{reads}{output of \link{assignReadingFrame}}

\item{tx_name}{transcript name.}

\item{col}{colors for reading frames}
}
\description{
Plot the bundances of P site on a transcript.
}
\examples{
pcs <- readRDS(system.file("extdata", "samplePc.rds",
               package="ribosomeProfilingQC"))

plotTranscript(pcs, c("ENSDART00000152562", "ENSDART00000054987"))
}
