% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCDS.R
\name{prepareCDS}
\alias{prepareCDS}
\title{prepare CDS}
\usage{
prepareCDS(txdb, withUTR = FALSE)
}
\arguments{
\item{txdb}{an TxDb object.}

\item{withUTR}{including UTR information or not.}
}
\value{
an GRanges object with metadata which include:
tx_id: transcript id;
tx_name: transcript name;
gene_id: gene id;
isFirstExonInCDS: is first exon in CDS or not;
idFirstExonInCDS: the id for the first exon;
isLastExonInCDS: is last exon in CDS or not;
wid.cumsu: cumulative sums of number of bases in CDS;
internalPos: offset position from 1 base;
}
\description{
prepare CDS library from a TxDb object.
}
\examples{
library(AnnotationDbi)
txdb <- loadDb(system.file("extdata", "danRer10.chr1.txdb",
               package="ribosomeProfilingQC"))
CDS <- prepareCDS(txdb)

}
