% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsDistribution.R
\name{readsDistribution}
\alias{readsDistribution}
\title{plot reads distribution in genomic elements}
\usage{
readsDistribution(
  reads,
  txdb,
  upstreamRegion = 3000,
  downstreamRegion = 3000,
  plot = TRUE,
  ...
)
}
\arguments{
\item{reads}{output of \link{getPsiteCoordinates}}

\item{txdb}{an TxDb object}

\item{upstreamRegion, downstreamRegion}{the range for promoter region and downstream region.}

\item{plot}{plot the distribution or not}

\item{...}{not use.}
}
\value{
the reads with distribution assignment
}
\description{
Plot the percentage of reads in CDS, 5'UTR, 3'UTR, introns, and other elements.
}
\examples{
library(Rsamtools)
bamfilename <- system.file("extdata", "RPF.chr1.bam",
                           package="ribosomeProfilingQC")
yieldSize <- 10000000
bamfile <- BamFile(bamfilename, yieldSize = yieldSize)
pc <- getPsiteCoordinates(bamfile, bestpsite=11)
pc.sub <- pc[pc$qwidth \%in\% c(29, 30)]
library(AnnotationDbi)
txdb <- loadDb(system.file("extdata", "danRer10.chr1.txdb",
               package="ribosomeProfilingQC"))
pc.sub <- readsDistribution(pc.sub, txdb, las=2)
}
