% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readsEndPlot.R
\name{readsEndPlot}
\alias{readsEndPlot}
\title{plot start/stop windows}
\usage{
readsEndPlot(
  bamfile,
  CDS,
  toStartCodon = TRUE,
  fiveEnd = TRUE,
  window = c(-29, 30),
  readLen = 25:30
)
}
\arguments{
\item{bamfile}{an BamFile object.}

\item{CDS}{output of \link{prepareCDS}}

\item{toStartCodon}{start or end codon}

\item{fiveEnd}{five or three ends of the reads}

\item{window}{the window of CDS region to plot}

\item{readLen}{the reads length used to plot}
}
\value{
the invisible counts numbers.
}
\description{
plot the reads shifted from start/stop position of CDS.
}
\examples{
library(Rsamtools)
bamfilename <- system.file("extdata", "RPF.WT.1.bam",
                           package="ribosomeProfilingQC")
yieldSize <- 10000000
bamfile <- BamFile(bamfilename, yieldSize = yieldSize)
#library(GenomicFeatures)
library(BSgenome.Drerio.UCSC.danRer10)
#txdb <- makeTxDbFromGFF(system.file("extdata",
 #         "Danio_rerio.GRCz10.91.chr1.gtf.gz",
 #         package="ribosomeProfilingQC"),
 #         organism = "Danio rerio",
 #         chrominfo = seqinfo(Drerio)["chr1"],
 #         taxonomyId = 7955)
#CDS <- prepareCDS(txdb)
CDS <- readRDS(system.file("extdata", "CDS.rds",
                           package="ribosomeProfilingQC"))
readsEndPlot(bamfile, CDS, toStartCodon=TRUE)
readsEndPlot(bamfile, CDS, toStartCodon=TRUE, fiveEnd=FALSE)
#readsEndPlot(bamfile, CDS, toStartCodon=FALSE)
#readsEndPlot(bamfile, CDS, toStartCodon=FALSE, fiveEnd=FALSE)
}
