% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftReadsByFrame.R
\name{shiftReadsByFrame}
\alias{shiftReadsByFrame}
\title{shift reads by reading frame}
\usage{
shiftReadsByFrame(reads, txdb)
}
\arguments{
\item{reads}{output of \link{getPsiteCoordinates}}

\item{txdb}{TxDb object.}
}
\value{
reads with reading frame information
}
\description{
shift reads P site position by reading frame. After shifting, all reading frame will be set as 0
}
\examples{
library(Rsamtools)
bamfilename <- system.file("extdata", "tophat2.danRer10.RPF.chr1.bam",
                           package="ribosomeProfilingQC")
yieldSize <- 10000000
bamfile <- BamFile(bamfilename, yieldSize = yieldSize)
pc <- getPsiteCoordinates(bamfile, bestpsite=11)
pc.sub <- pc[pc$qwidth \%in\% c(29, 30)]
library(AnnotationDbi)
txdb <- loadDb(system.file("extdata", "danRer10.chr1.txdb",
               package="ribosomeProfilingQC"))
pc.sub <- shiftReadsByFrame(pc.sub, txdb)

}
