% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandPlot.R
\name{strandPlot}
\alias{strandPlot}
\title{plot the distribution of reads in sense and antisense strand}
\usage{
strandPlot(reads, CDS, col = c("#009E73", "#D55E00"), ...)
}
\arguments{
\item{reads}{output of \link{getPsiteCoordinates}}

\item{CDS}{output of \link{prepareCDS}}

\item{col}{coloar for sense and antisense strand.}

\item{...}{parameter passed to barplot}
}
\value{
ggplot object.
}
\description{
plot the distribution of reads in sense and antisense strand to
check the mapping is correct.
}
\examples{
library(Rsamtools)
bamfilename <- system.file("extdata", "tophat2.danRer10.RPF.chr1.bam",
                           package="ribosomeProfilingQC")
yieldSize <- 10000000
bamfile <- BamFile(bamfilename, yieldSize = yieldSize)
pc <- getPsiteCoordinates(bamfile, bestpsite=11)
pc.sub <- pc[pc$qwidth \%in\% c(29, 30)]
CDS <- readRDS(system.file("extdata", "sampleCDS.rds",
               package="ribosomeProfilingQC"))
strandPlot(pc.sub, CDS)

}
